+++ R/warbleR-package.R
@@ -161,6 +161,7 @@
 #' @import parallel
 #' @importFrom methods formalArgs new is slotNames
 #' @importFrom dtw dtwDist
+#' @importFrom Rcpp evalCpp
 #' @importFrom stats cor dist aggregate approx ave prcomp time ts predict smooth.spline complete.cases spline weighted.mean na.omit var sd rlnorm fft
 #' 
 #' @author Marcelo Araya-Salas & Grace Smith Vidaurre